#2858: make moreh_helper_functions.hpp
