+++ tt_eager/tt_dnn/op_library/moreh_softmax/moreh_softmax_op.cpp
@@ -3,8 +3,8 @@
 // SPDX-License-Identifier: Apache-2.0
 
 #include "tt_eager/tt_dnn/op_library/moreh_softmax/moreh_softmax_op.hpp"
+#include "tt_eager/tt_dnn/op_library/moreh_helper_functions.hpp"
 #include "tt_eager/tt_dnn/op_library/work_split.hpp"
-#include "tt_eager/tt_dnn/op_library/moreh_softmax/helper_functions.hpp"
 #include "tt_dnn/op_library/run_operation.hpp"
 
 #include "tt_metal/host_api.hpp"
